requirements.txt
Django==3.2
graphene-django 2.15.0 
django-graphql-jwt==0.3.4
